# swagger/api.yaml
openapi: 3.0.3
info:
  title: Takween NestJS API
  description: |
    Complete API documentation for Takween NestJS application.

    This API provides endpoints for managing:
    - User authentication and registration (phone verification)
    - Entity management
    - Activities
    - Classrooms
    - Students

    **Authentication**: Most endpoints require Bearer token authentication. Use the `/user/phone/verify` endpoint to obtain an access token. User authentication uses phone number verification with SMS codes.
  version: 1.0.0
  contact:
    name: Takween API Support

servers:
  - url: http://127.0.0.1:3000
    description: Local development server
  - url: https://nestjs-api-dev-e8b7t.ondigitalocean.app
    description: development server

tags:
  - name: Authentication
    description: Token management and refresh
  - name: User
    description: User registration and phone verification
  - name: Entity
    description: Entity management (requires authentication)
  - name: Activity
    description: Activity management
  - name: Classroom
    description: Classroom management
  - name: Student
    description: Student management

paths:
  # Authentication Endpoints
  /token/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get a new access token using a refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
                  description: Valid refresh token
            example:
              refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '201':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    description: New JWT access token
                  refreshToken:
                    type: string
                    description: New JWT refresh token
              example:
                accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          $ref: '#/components/responses/Unauthorized'

  # User Endpoints
  /user/phone/register:
    post:
      tags:
        - User
      summary: Register user and send verification code
      description: Register a new user account (admin, teacher, or parent) and send a verification code to the phone number
      operationId: phoneRegisterUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhoneRegisterDto'
            example:
              phone: '1234567890'
              type: admin
      responses:
        '201':
          description: User registered and verification code sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          description: Too many requests - Please wait before requesting a new code

  /user/phone/verify:
    post:
      tags:
        - User
      summary: Verify phone number and login
      description: Verify the phone verification code and receive access and refresh tokens
      operationId: phoneVerifyUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhoneVerifyDto'
            example:
              phone: '1234567890'
              code: '123456'
      responses:
        '200':
          description: Phone verified and login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  admin:
                    type: object
                    properties:
                      id:
                        type: string
                        description: MongoDB ObjectId
                        example: 507f1f77bcf86cd799439011
                      name:
                        type: string
                        example: John Doe
                      phone:
                        type: string
                        example: '1234567890'
                      accessToken:
                        type: string
                        description: JWT access token (Bearer token)
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                      refreshToken:
                        type: string
                        description: JWT refresh token
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
              example:
                admin:
                  id: 507f1f77bcf86cd799439011
                  name: John Doe
                  phone: '1234567890'
                  accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /user:
    get:
      tags:
        - User
      summary: Get current user
      description: Get information about the authenticated user
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - User
      summary: Update current user
      description: Update the authenticated user's information
      operationId: updateCurrentUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
            example:
              name: John Updated
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - User
      summary: Delete current user
      description: Delete the authenticated user account
      operationId: deleteCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: تم حذف المستخدم بنجاح
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # Entity Endpoints (Requires Admin Authentication)
  /entity:
    get:
      tags:
        - Entity
      summary: Get all entities
      description: Retrieve a list of all entities for the authenticated admin (requires admin authentication)
      operationId: getEntities
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of entities retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Entity'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Forbidden - Admin access required

    post:
      tags:
        - Entity
      summary: Create entity
      description: Create a new entity (requires admin authentication)
      operationId: createEntity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEntityDto'
            example:
              name: School District
      responses:
        '201':
          description: Entity created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Forbidden - Admin access required
        '404':
          $ref: '#/components/responses/NotFound'

  /entity/{id}:
    put:
      tags:
        - Entity
      summary: Update entity
      description: Update an entity by ID (requires admin authentication)
      operationId: updateEntity
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEntityDto'
            example:
              name: Updated Entity Name
      responses:
        '200':
          description: Entity updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Forbidden - Admin access required
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Entity
      summary: Delete entity
      description: Delete an entity by ID (requires admin authentication)
      operationId: deleteEntity
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Entity deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: تم حذف الكيان بنجاح
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Forbidden - Admin access required
        '404':
          $ref: '#/components/responses/NotFound'

  # Activity Endpoints (Requires Admin Authentication)
  /activity:
    get:
      tags:
        - Activity
      summary: Get all activities
      description: Retrieve a list of all activities (requires admin authentication)
      operationId: getActivities
      security:
        - bearerAuth: []
      parameters:
        - name: entity
          in: query
          required: false
          description: Filter activities by entity ID
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: List of activities retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Activity'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Forbidden - Admin access required

    post:
      tags:
        - Activity
      summary: Create activity
      description: Create a new activity (requires admin authentication)
      operationId: createActivity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateActivityDto'
            example:
              name: Mathematics Course
              entity: 507f1f77bcf86cd799439011
              price: 100
      responses:
        '201':
          description: Activity created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Forbidden - Admin access required
        '404':
          $ref: '#/components/responses/NotFound'

  /activity/{id}:
    put:
      tags:
        - Activity
      summary: Update activity
      description: Update an activity by ID (requires admin authentication)
      operationId: updateActivity
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateActivityDto'
            example:
              name: Updated Activity Name
              price: 150
      responses:
        '200':
          description: Activity updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Forbidden - Admin access required
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Activity
      summary: Delete activity
      description: Delete an activity by ID (requires admin authentication)
      operationId: deleteActivity
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Activity deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: تم حذف النشاط بنجاح
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Forbidden - Admin access required
        '404':
          $ref: '#/components/responses/NotFound'

  # Classroom Endpoints
  /classroom:
    get:
      tags:
        - Classroom
      summary: Get all classrooms
      description: Retrieve a list of all classrooms
      operationId: getClassrooms
      responses:
        '200':
          description: List of classrooms retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Classroom'

    post:
      tags:
        - Classroom
      summary: Create classroom
      description: Create a new classroom
      operationId: createClassroom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClassroomDto'
            example:
              name: Grade 5A
              activity: 507f1f77bcf86cd799439011
      responses:
        '201':
          description: Classroom created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Classroom'
        '400':
          $ref: '#/components/responses/BadRequest'

  /classroom/{id}:
    put:
      tags:
        - Classroom
      summary: Update classroom
      description: Update a classroom by ID
      operationId: updateClassroom
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClassroomDto'
            example:
              name: Grade 5B
      responses:
        '200':
          description: Classroom updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Classroom'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Classroom
      summary: Delete classroom
      description: Delete a classroom by ID
      operationId: deleteClassroom
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Classroom deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'

  # Student Endpoints
  /student:
    get:
      tags:
        - Student
      summary: Get all students
      description: Retrieve a list of all students
      operationId: getStudents
      responses:
        '200':
          description: List of students retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'

    post:
      tags:
        - Student
      summary: Create student
      description: Create a new student
      operationId: createStudent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStudentDto'
            example:
              name: Jane Smith
              email: jane.smith@example.com
              grade: 5
              age: 10
      responses:
        '201':
          description: Student created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          $ref: '#/components/responses/BadRequest'

  /student/{id}:
    get:
      tags:
        - Student
      summary: Get student by ID
      description: Retrieve a specific student by ID
      operationId: getStudent
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Student retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Student
      summary: Update student
      description: Update a student by ID
      operationId: updateStudent
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStudentDto'
            example:
              name: Jane Updated
              email: jane.updated@example.com
              grade: 6
              age: 11
      responses:
        '200':
          description: Student updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Student
      summary: Delete student
      description: Delete a student by ID
      operationId: deleteStudent
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Student deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'


components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT authentication using Bearer token.
        Format: `Authorization: Bearer <token>`
        Obtain token from `/user/phone/verify` endpoint.

  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      description: MongoDB ObjectId
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
      example: 507f1f77bcf86cd799439011

  schemas:
    # Entity DTOs
    CreateEntityDto:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Entity name
          example: School District
        owner:
          type: string
          pattern: '^[0-9a-fA-F]{24}$'
          description: MongoDB ObjectId of the user owner (optional, defaults to authenticated user)
          example: 507f1f77bcf86cd799439011

    UpdateEntityDto:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Entity name
          example: Updated Entity Name

    Entity:
      type: object
      properties:
        _id:
          type: string
          description: MongoDB ObjectId
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          example: School District
        owner:
          type: string
          description: MongoDB ObjectId of the user owner
          example: 507f1f77bcf86cd799439011
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Activity DTOs
    CreateActivityDto:
      type: object
      required:
        - name
        - entity
      properties:
        name:
          type: string
          description: Activity name
          example: Mathematics Course
        entity:
          type: string
          pattern: '^[0-9a-fA-F]{24}$'
          description: MongoDB ObjectId of the associated entity
          example: 507f1f77bcf86cd799439011
        price:
          type: number
          minimum: 0
          description: Activity price (optional, defaults to 0)
          example: 100

    UpdateActivityDto:
      type: object
      properties:
        name:
          type: string
          description: Activity name
          example: Updated Activity Name
        price:
          type: number
          minimum: 0
          description: Activity price (optional)
          example: 150

    Activity:
      type: object
      properties:
        _id:
          type: string
          description: MongoDB ObjectId
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          example: Mathematics Course
        entity:
          type: string
          example: 507f1f77bcf86cd799439011
        price:
          type: number
          description: Activity price
          example: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Classroom DTOs
    CreateClassroomDto:
      type: object
      required:
        - name
        - activity
      properties:
        name:
          type: string
          description: Classroom name
          example: Grade 5A
        activity:
          type: string
          pattern: '^[0-9a-fA-F]{24}$'
          description: MongoDB ObjectId of the associated activity
          example: 507f1f77bcf86cd799439011

    UpdateClassroomDto:
      type: object
      properties:
        name:
          type: string
          description: Classroom name
          example: Grade 5B

    Classroom:
      type: object
      properties:
        _id:
          type: string
          description: MongoDB ObjectId
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          example: Grade 5A
        activity:
          type: string
          example: 507f1f77bcf86cd799439011
        teachers:
          type: array
          items:
            type: string
          description: Array of teacher MongoDB ObjectIds
          example:
            - 507f1f77bcf86cd799439011
        students:
          type: array
          items:
            type: string
          description: Array of student MongoDB ObjectIds
          example:
            - 507f1f77bcf86cd799439012
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Student DTOs
    CreateStudentDto:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          description: Student full name
          example: Jane Smith
        email:
          type: string
          format: email
          description: Student email address
          example: jane.smith@example.com
        grade:
          type: string
          description: Student grade level
          example: '5'
        age:
          type: integer
          minimum: 0
          maximum: 120
          description: Student age
          example: 10

    UpdateStudentDto:
      type: object
      properties:
        name:
          type: string
          description: Student full name
          example: Jane Updated
        email:
          type: string
          format: email
          description: Student email address
          example: jane.updated@example.com
        grade:
          type: string
          description: Student grade level
          example: '6'
        age:
          type: integer
          minimum: 0
          maximum: 120
          description: Student age
          example: 11

    Student:
      type: object
      properties:
        _id:
          type: string
          description: MongoDB ObjectId
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          example: Jane Smith
        email:
          type: string
          format: email
          example: jane.smith@example.com
        grade:
          type: string
          example: '5'
        age:
          type: integer
          example: 10
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # User DTOs
    PhoneRegisterDto:
      type: object
      required:
        - phone
        - type
      properties:
        phone:
          type: string
          minLength: 7
          maxLength: 15
          description: Phone number (numeric string)
          example: '1234567890'
        type:
          type: string
          enum:
            - admin
            - teacher
            - parent
          description: User type
          example: admin

    PhoneVerifyDto:
      type: object
      required:
        - phone
        - code
      properties:
        phone:
          type: string
          minLength: 7
          maxLength: 15
          description: Phone number (numeric string)
          example: '1234567890'
        code:
          type: string
          description: Verification code sent to phone
          example: '123456'

    UpdateUserDto:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 50
          description: User full name (optional)
          example: John Updated

    # User Model
    User:
      type: object
      properties:
        _id:
          type: string
          description: MongoDB ObjectId
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          minLength: 3
          maxLength: 50
          description: User full name (optional)
          example: John Doe
        phone:
          type: string
          minLength: 7
          maxLength: 15
          description: Phone number (unique)
          example: '1234567890'
        type:
          type: string
          enum:
            - admin
            - teacher
            - parent
          description: User type
          example: admin
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

  responses:
    BadRequest:
      description: Bad request - Invalid input data
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                example: 400
              message:
                type: array
                items:
                  type: string
                example:
                  - name must be longer than or equal to 3 characters
                  - phone must be a number string
              error:
                type: string
                example: Bad Request

    Unauthorized:
      description: Unauthorized - Missing or invalid authentication token
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                example: 401
              message:
                type: string
                example: Invalid or expired access token
              error:
                type: string
                example: Unauthorized

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                example: 404
              message:
                type: string
                example: Entity not found
              error:
                type: string
                example: Not Found